Change to welcome message; something for myself.
The welcome message link to #rules (a channel in my server) has been updated to actually link to said channel. Additionally, given that this bot is mainly being developed for use in my server, I have made a small change that should work so much better to cater to rules there and also allow him to be 100% functional in other servers with no hassle.