Complete overhaul of Yoshi-Bot's command reading.
I decided to take the time to completely rehaul everything about Yoshi's
method in finding what command to use. I transferred all commands to a
new file called "commands.js," which is practically just an object
holding each command name as an attribute, consequently having
attributes 'usage,' 'description,' and 'process,' which hold an example
of usage of the command, a description of what the command does, and the
function it must run, respectively. I do this both to get away from
clunky conditional statements, and instead organize everything into much
more readable and debuggable code. Separate files are good, so long as
things look organized, which is what I aim to do now. Anyhow, that's it
for this! Commenting my code is next, so look forward to that. Cheers.